ai_driver/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements/requirements.in
#
aiohttp==3.8.5
    # via
    #   datasets
    #   fsspec
    #   langchain
    #   openai
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.5.0
    # via pydantic
anyio==3.7.1
    # via
    #   -r requirements/requirements.in
    #   httpcore
    #   starlette
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==23.1.0
    # via
    #   aiohttp
    #   jsonlines
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.2.0
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via
    #   nltk
    #   uvicorn
cmake==3.27.2
    # via triton
ctransformers==0.2.22
    # via -r requirements/requirements.in
dataclasses-json==0.5.14
    # via langchain
datasets==2.14.4
    # via -r requirements/requirements.in
dill==0.3.7
    # via
    #   datasets
    #   multiprocess
dnspython==2.4.2
    # via pinecone-client
exceptiongroup==1.1.3
    # via anyio
faiss-cpu==1.7.4
    # via -r requirements/requirements.in
fastapi==0.101.1
    # via -r requirements/requirements.in
filelock==3.12.2
    # via
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2023.6.0
    # via
    #   datasets
    #   huggingface-hub
googleapis-common-protos==1.60.0
    # via
    #   grpc-gateway-protoc-gen-openapiv2
    #   pinecone-client
greenlet==2.0.2
    # via sqlalchemy
grpc-gateway-protoc-gen-openapiv2==0.1.0
    # via pinecone-client
grpcio==1.57.0
    # via pinecone-client
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.17.3
    # via httpx
httpx==0.24.1
    # via -r requirements/requirements.in
huggingface-hub==0.16.4
    # via
    #   ctransformers
    #   datasets
    #   sentence-transformers
    #   transformers
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
instructorembedding==1.0.1
    # via -r requirements/requirements.in
jinja2==3.1.2
    # via torch
joblib==1.3.2
    # via
    #   nltk
    #   scikit-learn
jsonlines==3.1.0
    # via -r requirements/requirements.in
langchain==0.0.268
    # via -r requirements/requirements.in
langsmith==0.0.25
    # via
    #   -r requirements/requirements.in
    #   langchain
lit==16.0.6
    # via triton
loguru==0.7.0
    # via
    #   -r requirements/requirements.in
    #   pinecone-client
lz4==4.3.2
    # via pinecone-client
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via jinja2
marshmallow==3.20.1
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.15
    # via datasets
mypy-extensions==1.0.0
    # via typing-inspect
networkx==3.1
    # via torch
nltk==3.8.1
    # via sentence-transformers
numexpr==2.8.5
    # via langchain
numpy==1.25.2
    # via
    #   -r requirements/requirements.in
    #   datasets
    #   langchain
    #   numexpr
    #   pandas
    #   pinecone-client
    #   pyarrow
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   torchvision
    #   transformers
nvidia-cublas-cu11==11.10.3.66
    # via
    #   nvidia-cudnn-cu11
    #   nvidia-cusolver-cu11
    #   torch
nvidia-cuda-cupti-cu11==11.7.101
    # via torch
nvidia-cuda-nvrtc-cu11==11.7.99
    # via torch
nvidia-cuda-runtime-cu11==11.7.99
    # via torch
nvidia-cudnn-cu11==8.5.0.96
    # via torch
nvidia-cufft-cu11==10.9.0.58
    # via torch
nvidia-curand-cu11==10.2.10.91
    # via torch
nvidia-cusolver-cu11==11.4.0.1
    # via torch
nvidia-cusparse-cu11==11.7.4.91
    # via torch
nvidia-nccl-cu11==2.14.3
    # via torch
nvidia-nvtx-cu11==11.7.91
    # via torch
openai==0.27.8
    # via -r requirements/requirements.in
packaging==23.1
    # via
    #   datasets
    #   huggingface-hub
    #   marshmallow
    #   transformers
pandas==2.0.3
    # via
    #   -r requirements/requirements.in
    #   datasets
pillow==10.0.0
    # via torchvision
pinecone-client[grpc]==2.2.2
    # via -r requirements/requirements.in
protobuf==3.19.6
    # via
    #   googleapis-common-protos
    #   pinecone-client
py-cpuinfo==9.0.0
    # via ctransformers
pyarrow==12.0.1
    # via
    #   -r requirements/requirements.in
    #   datasets
pydantic==2.2.1
    # via
    #   fastapi
    #   langchain
    #   langsmith
pydantic-core==2.6.1
    # via pydantic
pygments==2.16.1
    # via rich
pypdf==3.15.1
    # via -r requirements/requirements.in
pypdf2==3.0.1
    # via -r requirements/requirements.in
python-box==7.0.1
    # via -r requirements/requirements.in
python-dateutil==2.8.2
    # via
    #   pandas
    #   pinecone-client
python-dotenv==1.0.0
    # via -r requirements/requirements.in
pytz==2023.3
    # via pandas
pyyaml==6.0.1
    # via
    #   datasets
    #   huggingface-hub
    #   langchain
    #   pinecone-client
    #   transformers
regex==2023.8.8
    # via
    #   nltk
    #   tiktoken
    #   transformers
requests==2.31.0
    # via
    #   -r requirements/requirements.in
    #   datasets
    #   fsspec
    #   huggingface-hub
    #   langchain
    #   langsmith
    #   openai
    #   pinecone-client
    #   tiktoken
    #   torchvision
    #   transformers
rich==13.5.2
    # via -r requirements/requirements.in
safetensors==0.3.2
    # via transformers
scikit-learn==1.3.0
    # via
    #   -r requirements/requirements.in
    #   sentence-transformers
scipy==1.11.2
    # via
    #   -r requirements/requirements.in
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==2.2.2
    # via -r requirements/requirements.in
sentencepiece==0.1.99
    # via sentence-transformers
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==2.0.20
    # via langchain
starlette==0.27.0
    # via fastapi
sympy==1.12
    # via torch
tdqm==0.0.1
    # via -r requirements/requirements.in
tenacity==8.2.3
    # via langchain
threadpoolctl==3.2.0
    # via scikit-learn
tiktoken==0.4.0
    # via -r requirements/requirements.in
tokenizers==0.13.3
    # via transformers
torch==2.0.1
    # via
    #   -r requirements/requirements.in
    #   sentence-transformers
    #   torchvision
    #   triton
torchvision==0.15.2
    # via sentence-transformers
tqdm==4.66.1
    # via
    #   datasets
    #   huggingface-hub
    #   nltk
    #   openai
    #   pinecone-client
    #   sentence-transformers
    #   tdqm
    #   transformers
transformers==4.31.0
    # via
    #   -r requirements/requirements.in
    #   sentence-transformers
triton==2.0.0
    # via torch
typing-extensions==4.7.1
    # via
    #   fastapi
    #   huggingface-hub
    #   pinecone-client
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   torch
    #   typing-inspect
    #   uvicorn
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2023.3
    # via pandas
urllib3==2.0.4
    # via
    #   pinecone-client
    #   requests
uvicorn==0.23.2
    # via -r requirements/requirements.in
wheel==0.41.1
    # via
    #   nvidia-cublas-cu11
    #   nvidia-cuda-cupti-cu11
    #   nvidia-cuda-runtime-cu11
    #   nvidia-curand-cu11
    #   nvidia-cusparse-cu11
    #   nvidia-nvtx-cu11
xxhash==3.3.0
    # via datasets
yarl==1.9.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
